TST: separate Fourier tests and parametrize more
