+++ odl/test/trafos/fourier_test.py
@@ -28,59 +28,15 @@ import numpy as np
 import pytest
 
 import odl
-from odl.trafos.backends import pyfftw_call
 from odl.trafos.util.ft_utils import (
-    reciprocal_grid, realspace_grid, dft_preprocess_data, dft_postprocess_data,
+    reciprocal_grid, dft_preprocess_data, dft_postprocess_data,
     _interp_kernel_ft)
 from odl.trafos.fourier import (
     DiscreteFourierTransform, DiscreteFourierTransformInverse,
     FourierTransform)
-from odl.util import (all_almost_equal, all_equal,
+from odl.util import (all_almost_equal,
                       never_skip, skip_if_no_pyfftw,
                       is_real_dtype, conj_exponent, TYPE_MAP_R2C)
-
-
-exp_params = [2.0, 1.0, float('inf'), 1.5]
-exp_ids = [' p = {} '.format(p) for p in exp_params]
-
-
-@pytest.fixture(scope="module", ids=exp_ids, params=exp_params)
-def exponent(request):
-    return request.param
-
-
-dtype_params = [str(dtype) for dtype in TYPE_MAP_R2C.keys()]
-dtype_params += [str(dtype) for dtype in TYPE_MAP_R2C.values()]
-dtype_params = list(set(dtype_params))
-dtype_ids = [' dtype = {} '.format(dt) for dt in dtype_params]
-
-
-@pytest.fixture(scope="module", ids=dtype_ids, params=dtype_params)
-def dtype(request):
-    return request.param
-
-
-plan_params = ['estimate', 'measure', 'patient', 'exhaustive']
-plan_ids = [" planning = '{}' ".format(p) for p in plan_params]
-
-
-@pytest.fixture(scope="module", ids=plan_ids, params=plan_params)
-def planning(request):
-    return request.param
-
-
-impl_params = [never_skip('numpy'), skip_if_no_pyfftw('pyfftw')]
-impl_ids = ['impl={}'.format(impl.args[1]) for impl in impl_params]
-
-
-@pytest.fixture(scope="module", ids=impl_ids, params=impl_params)
-def impl(request):
-    return request.param
-
-
-@pytest.fixture(scope='module', ids=[' - ', ' + '], params=['-', '+'])
-def sign(request):
-    return request.param
 
 
 def _random_array(shape, dtype):
@@ -91,720 +47,16 @@ def _random_array(shape, dtype):
                 1j * np.random.rand(*shape).astype(dtype))
 
 
-@pytest.fixture(scope='module', ids=[' forward ', ' backward '],
-                params=['forward', 'backward'])
-def direction(request):
-    return request.param
-
-
-# ---- reciprocal_grid ---- #
-
-
-def test_reciprocal_grid_1d_odd():
-
-    grid = odl.uniform_sampling(0, 1, shape=11)
-    s = grid.stride
-    n = np.array(grid.shape)
-
-    true_recip_stride = 2 * pi / (s * n)
-
-    # Without shift
-    rgrid = reciprocal_grid(grid, shift=False, halfcomplex=False)
-
-    # Independent of shift and halfcomplex, check anyway
-    assert all_equal(rgrid.shape, n)
-    assert all_almost_equal(rgrid.stride, true_recip_stride)
-    # Should be symmetric
-    assert all_almost_equal(rgrid.min_pt, -rgrid.max_pt)
-    assert all_almost_equal(rgrid.mid_pt, 0)
-    # Zero should be at index n // 2
-    assert all_almost_equal(rgrid[n // 2], 0)
-
-    # With shift
-    rgrid = reciprocal_grid(grid, shift=True, halfcomplex=False)
-
-    # Independent of shift and halfcomplex, check anyway
-    assert all_equal(rgrid.shape, n)
-    assert all_almost_equal(rgrid.stride, true_recip_stride)
-    # No point should be closer to 0 than half a recip stride
-    atol = 0.999 * true_recip_stride / 2
-    assert not rgrid.approx_contains(0, atol=atol)
-
-    # Inverting should give back the original
-    irgrid = realspace_grid(rgrid, grid.min_pt, halfcomplex=False)
-    assert irgrid.approx_equals(grid, atol=1e-6)
-
-
-def test_reciprocal_1d_odd_halfcomplex():
-
-    grid = odl.uniform_sampling(0, 1, shape=11)
-    s = grid.stride
-    n = np.array(grid.shape)
-
-    true_recip_stride = 2 * pi / (s * n)
-
-    # Without shift
-    rgrid = reciprocal_grid(grid, shift=False, halfcomplex=True)
-
-    # Independent of shift and halfcomplex, check anyway
-    assert all_equal(rgrid.shape, (n + 1) / 2)
-    assert all_almost_equal(rgrid.stride, true_recip_stride)
-    # Max should be zero
-    assert all_almost_equal(rgrid.max_pt, 0)
-
-    # With shift
-    rgrid = reciprocal_grid(grid, shift=True, halfcomplex=True)
-
-    # Independent of shift and halfcomplex, check anyway
-    assert all_equal(rgrid.shape, (n + 1) / 2)
-    assert all_almost_equal(rgrid.stride, true_recip_stride)
-    # Max point should be half a positive recip stride
-    assert all_almost_equal(rgrid.max_pt, -true_recip_stride / 2)
-
-    # Inverting should give back the original
-    irgrid = realspace_grid(rgrid, grid.min_pt, halfcomplex=True,
-                            halfcx_parity='odd')
-    assert irgrid.approx_equals(grid, atol=1e-6)
-
-
-def test_reciprocal_grid_1d_even():
-
-    grid = odl.uniform_sampling(0, 1, shape=10)
-    s = grid.stride
-    n = np.array(grid.shape)
-
-    true_recip_stride = 2 * pi / (s * n)
-
-    # Without shift
-    rgrid = reciprocal_grid(grid, shift=False, halfcomplex=False)
-
-    # Independent of shift and halfcomplex, check anyway
-    assert all_equal(rgrid.shape, n)
-    assert all_almost_equal(rgrid.stride, true_recip_stride)
-    # Should be symmetric
-    assert all_almost_equal(rgrid.min_pt, -rgrid.max_pt)
-    assert all_almost_equal(rgrid.mid_pt, 0)
-    # No point should be closer to 0 than half a recip stride
-    atol = 0.999 * true_recip_stride / 2
-    assert not rgrid.approx_contains(0, atol=atol)
-
-    # With shift
-    rgrid = reciprocal_grid(grid, shift=True, halfcomplex=False)
-
-    # Independent of shift and halfcomplex, check anyway
-    assert all_equal(rgrid.shape, n)
-    assert all_almost_equal(rgrid.stride, true_recip_stride)
-    # Zero should be at index n // 2
-    assert all_almost_equal(rgrid[n // 2], 0)
-
-    # Inverting should give back the original
-    irgrid = realspace_grid(rgrid, grid.min_pt, halfcomplex=False)
-    assert irgrid.approx_equals(grid, atol=1e-6)
-
-
-def test_reciprocal_grid_1d_even_halfcomplex():
-
-    grid = odl.uniform_sampling(0, 1, shape=10)
-    s = grid.stride
-    n = np.array(grid.shape)
-
-    true_recip_stride = 2 * pi / (s * n)
-
-    # Without shift
-    rgrid = reciprocal_grid(grid, shift=False, halfcomplex=True)
-
-    # Independent of shift and halfcomplex, check anyway
-    assert all_equal(rgrid.shape, n / 2 + 1)
-    assert all_almost_equal(rgrid.stride, true_recip_stride)
-    # Max point should be half a positive recip stride
-    assert all_almost_equal(rgrid.max_pt, true_recip_stride / 2)
-
-    # With shift
-    rgrid = reciprocal_grid(grid, shift=True, halfcomplex=True)
-
-    # Independent of shift and halfcomplex, check anyway
-    assert all_equal(rgrid.shape, n / 2 + 1)
-    assert all_almost_equal(rgrid.stride, true_recip_stride)
-    # Max should be zero
-    assert all_almost_equal(rgrid.max_pt, 0)
-
-    # Inverting should give back the original
-    irgrid = realspace_grid(rgrid, grid.min_pt, halfcomplex=True,
-                            halfcx_parity='even')
-    assert irgrid.approx_equals(grid, atol=1e-6)
-
-
-def test_reciprocal_grid_nd():
-
-    grid = odl.uniform_sampling([0] * 3, [1] * 3, shape=(3, 4, 5))
-    s = grid.stride
-    n = np.array(grid.shape)
-
-    true_recip_stride = 2 * pi / (s * n)
-
-    # Without shift altogether
-    rgrid = reciprocal_grid(grid, shift=False, halfcomplex=False)
-
-    assert all_equal(rgrid.shape, n)
-    assert all_almost_equal(rgrid.stride, true_recip_stride)
-    assert all_almost_equal(rgrid.min_pt, -rgrid.max_pt)
-
-    # Inverting should give back the original
-    irgrid = realspace_grid(rgrid, grid.min_pt, halfcomplex=False)
-    assert irgrid.approx_equals(grid, atol=1e-6)
-
-
-def test_reciprocal_grid_nd_shift_list():
-
-    grid = odl.uniform_sampling([0] * 3, [1] * 3, shape=(3, 4, 5))
-    s = grid.stride
-    n = np.array(grid.shape)
-    shift = [False, True, False]
-
-    true_recip_stride = 2 * pi / (s * n)
-
-    # Shift only the even dimension, then zero must be contained
-    rgrid = reciprocal_grid(grid, shift=shift, halfcomplex=False)
-    noshift = np.where(np.logical_not(shift))
-
-    assert all_equal(rgrid.shape, n)
-    assert all_almost_equal(rgrid.stride, true_recip_stride)
-    assert all_almost_equal(rgrid.min_pt[noshift], -rgrid.max_pt[noshift])
-    assert all_almost_equal(rgrid[n // 2], [0] * 3)
-
-    # Inverting should give back the original
-    irgrid = realspace_grid(rgrid, grid.min_pt, halfcomplex=False)
-    assert irgrid.approx_equals(grid, atol=1e-6)
-
-
-def test_reciprocal_grid_nd_axes():
-
-    grid = odl.uniform_sampling([0] * 3, [1] * 3, shape=(3, 4, 5))
-    s = grid.stride
-    n = np.array(grid.shape)
-    axes_list = [[1, -1], [0], 0, [0, 2, 1], [2, 0]]
-
-    for axes in axes_list:
-        active = np.zeros(grid.ndim, dtype=bool)
-        active[axes] = True
-        inactive = np.logical_not(active)
-
-        true_recip_stride = np.empty(grid.ndim)
-        true_recip_stride[active] = 2 * pi / (s[active] * n[active])
-        true_recip_stride[inactive] = s[inactive]
-
-        # Without shift altogether
-        rgrid = reciprocal_grid(grid, shift=False, axes=axes,
-                                halfcomplex=False)
-
-        assert all_equal(rgrid.shape, n)
-        assert all_almost_equal(rgrid.stride, true_recip_stride)
-        assert all_almost_equal(rgrid.min_pt[active], -rgrid.max_pt[active])
-        assert all_equal(rgrid.min_pt[inactive], grid.min_pt[inactive])
-        assert all_equal(rgrid.max_pt[inactive], grid.max_pt[inactive])
-
-        # Inverting should give back the original
-        irgrid = realspace_grid(rgrid, grid.min_pt, axes=axes,
-                                halfcomplex=False)
-        assert irgrid.approx_equals(grid, atol=1e-6)
-
-
-def test_reciprocal_grid_nd_halfcomplex():
-
-    grid = odl.uniform_sampling([0] * 3, [1] * 3, shape=(3, 4, 5))
-    s = grid.stride
-    n = np.array(grid.shape)
-    stride_last = 2 * pi / (s[-1] * n[-1])
-    n[-1] = n[-1] // 2 + 1
-
-    # Without shift
-    rgrid = reciprocal_grid(grid, shift=False, halfcomplex=True)
-    assert all_equal(rgrid.shape, n)
-    assert rgrid.max_pt[-1] == 0  # last dim is odd
-
-    # With shift
-    rgrid = reciprocal_grid(grid, shift=True, halfcomplex=True)
-    assert all_equal(rgrid.shape, n)
-    assert rgrid.max_pt[-1] == -stride_last / 2
-
-    # Inverting should give back the original
-    irgrid = realspace_grid(rgrid, grid.min_pt, halfcomplex=True,
-                            halfcx_parity='odd')
-    assert irgrid.approx_equals(grid, atol=1e-6)
-
-    with pytest.raises(ValueError):
-        realspace_grid(rgrid, grid.min_pt, halfcomplex=True,
-                       halfcx_parity='+')
-
-# ---- dft_preprocess_data ---- #
-
-
-def test_dft_preprocess_data(sign):
-
-    shape = (2, 3, 4)
-
-    # With shift
-    correct_arr = []
-    for i, j, k in product(range(shape[0]), range(shape[1]), range(shape[2])):
-        correct_arr.append((1 + 1j) * (1 - 2 * ((i + j + k) % 2)))
-
-    arr = np.ones(shape, dtype='complex64') * (1 + 1j)
-    preproc = dft_preprocess_data(arr, shift=True, sign=sign)  # out-of-place
-    dft_preprocess_data(arr, shift=True, out=arr, sign=sign)  # in-place
-
-    assert all_almost_equal(preproc.ravel(), correct_arr)
-    assert all_almost_equal(arr.ravel(), correct_arr)
-
-    # Without shift
-    imag = 1j if sign == '-' else -1j
-    correct_arr = []
-    for i, j, k in product(range(shape[0]), range(shape[1]), range(shape[2])):
-        argsum = sum((idx * (1 - 1 / shp))
-                     for idx, shp in zip((i, j, k), shape))
-
-        correct_arr.append((1 + 1j) * np.exp(imag * np.pi * argsum))
-
-    arr = np.ones(shape, dtype='complex64') * (1 + 1j)
-    dft_preprocess_data(arr, shift=False, out=arr, sign=sign)
-
-    assert all_almost_equal(arr.ravel(), correct_arr)
-
-    # Bad input
-    with pytest.raises(ValueError):
-        dft_preprocess_data(arr, out=arr, sign=1)
-
-    arr = np.zeros(shape, dtype='S2')
-    with pytest.raises(ValueError):
-        dft_preprocess_data(arr)
-
-
-def test_dft_preprocess_data_halfcomplex(sign):
-
-    shape = (2, 3, 4)
-
-    # With shift
-    correct_arr = []
-    for i, j, k in product(range(shape[0]), range(shape[1]), range(shape[2])):
-        correct_arr.append(1 - 2 * ((i + j + k) % 2))
-
-    arr = np.ones(shape, dtype='float64')
-    preproc = dft_preprocess_data(arr, shift=True, sign=sign)  # out-of-place
-    out = np.empty_like(arr)
-    dft_preprocess_data(arr, shift=True, out=out, sign=sign)  # in-place
-    dft_preprocess_data(arr, shift=True, out=arr, sign=sign)  # in-place
-    assert all_almost_equal(preproc.ravel(), correct_arr)
-    assert all_almost_equal(arr.ravel(), correct_arr)
-    assert all_almost_equal(out.ravel(), correct_arr)
-
-    # Without shift
-    imag = 1j if sign == '-' else -1j
-    correct_arr = []
-    for i, j, k in product(range(shape[0]), range(shape[1]), range(shape[2])):
-        argsum = sum((idx * (1 - 1 / shp))
-                     for idx, shp in zip((i, j, k), shape))
-
-        correct_arr.append(np.exp(imag * np.pi * argsum))
-
-    arr = np.ones(shape, dtype='float64')
-    preproc = dft_preprocess_data(arr, shift=False, sign=sign)
-    assert all_almost_equal(preproc.ravel(), correct_arr)
-
-    # Non-float input works, too
-    arr = np.ones(shape, dtype='int')
-    preproc = dft_preprocess_data(arr, shift=False, sign=sign)
-    assert all_almost_equal(preproc.ravel(), correct_arr)
-
-    # In-place modification not possible for float array and no shift
-    arr = np.ones(shape, dtype='float64')
-    with pytest.raises(ValueError):
-        dft_preprocess_data(arr, shift=False, out=arr, sign=sign)
-
-
-def test_dft_preprocess_data_with_axes(sign):
-
-    shape = (2, 3, 4)
-
-    axes = 1  # Only middle index counts
-    correct_arr = []
-    for _, j, __ in product(range(shape[0]), range(shape[1]), range(shape[2])):
-        correct_arr.append(1 - 2 * (j % 2))
-
-    arr = np.ones(shape, dtype='complex64')
-    dft_preprocess_data(arr, shift=True, axes=axes, out=arr, sign=sign)
-
-    assert all_almost_equal(arr.ravel(), correct_arr)
-
-    axes = [0, -1]  # First and last
-    correct_arr = []
-    for i, _, k in product(range(shape[0]), range(shape[1]), range(shape[2])):
-        correct_arr.append(1 - 2 * ((i + k) % 2))
-
-    arr = np.ones(shape, dtype='complex64')
-    dft_preprocess_data(arr, shift=True, axes=axes, out=arr, sign=sign)
-
-    assert all_almost_equal(arr.ravel(), correct_arr)
-
-
-# ---- pyfftw_call ---- #
-
-
-def _params_from_dtype(dt):
-    if is_real_dtype(dt):
-        halfcomplex = True
-        dtype = TYPE_MAP_R2C[np.dtype(dt)]
-    else:
-        halfcomplex = False
-        dtype = dt
-    return halfcomplex, dtype
-
-
-def _halfcomplex_shape(shape, axes=None):
-    if axes is None:
-        axes = tuple(range(len(shape)))
-
-    try:
-        axes = (int(axes),)
-    except TypeError:
-        pass
-
-    shape = list(shape)
-    shape[axes[-1]] = shape[axes[-1]] // 2 + 1
-    return shape
-
-
-@skip_if_no_pyfftw
-def test_pyfftw_call_forward(dtype):
-    # Test against Numpy's FFT
-    if dtype == np.dtype('float16'):  # not supported, skipping
-        return
-
-    halfcomplex, out_dtype = _params_from_dtype(dtype)
-
-    for shape in [(10,), (3, 4, 5)]:
-        arr = _random_array(shape, dtype)
-
-        if halfcomplex:
-            true_dft = np.fft.rfftn(arr)
-            dft_arr = np.empty(_halfcomplex_shape(shape), dtype=out_dtype)
-        else:
-            true_dft = np.fft.fftn(arr)
-            dft_arr = np.empty(shape, dtype=out_dtype)
-
-        pyfftw_call(arr, dft_arr, direction='forward',
-                    halfcomplex=halfcomplex, preserve_input=False)
-
-        assert all_almost_equal(dft_arr, true_dft)
-
-
-@skip_if_no_pyfftw
-def test_pyfftw_call_threads():
-    shape = (3, 4, 5)
-    arr = _random_array(shape, dtype='complex64')
-    true_dft = np.fft.fftn(arr)
-    dft_arr = np.empty(shape, dtype='complex64')
-    pyfftw_call(arr, dft_arr, direction='forward', preserve_input=False,
-                threads=4)
-    assert all_almost_equal(dft_arr, true_dft)
-
-    shape = (1000,)  # Trigger cpu_count() as number of threads
-    arr = _random_array(shape, dtype='complex64')
-    true_dft = np.fft.fftn(arr)
-    dft_arr = np.empty(shape, dtype='complex64')
-    pyfftw_call(arr, dft_arr, direction='forward', preserve_input=False)
-
-    assert all_almost_equal(dft_arr, true_dft)
-
-
-@skip_if_no_pyfftw
-def test_pyfftw_call_backward(dtype):
-    # Test against Numpy's IFFT, no normalization
-    if dtype == np.dtype('float16'):  # not supported, skipping
-        return
-
-    halfcomplex, in_dtype = _params_from_dtype(dtype)
-
-    for shape in [(10,), (3, 4, 5)]:
-        # Scaling happens wrt output (large) shape
-        idft_scaling = np.prod(shape)
-
-        if halfcomplex:
-            arr = _random_array(_halfcomplex_shape(shape), in_dtype)
-            true_idft = np.fft.irfftn(arr, shape) * idft_scaling
-        else:
-            arr = _random_array(shape, in_dtype)
-            true_idft = np.fft.ifftn(arr) * idft_scaling
-
-        idft_arr = np.empty(shape, dtype=dtype)
-        pyfftw_call(arr, idft_arr, direction='backward',
-                    halfcomplex=halfcomplex)
-
-        assert all_almost_equal(idft_arr, true_idft)
-
-
-@skip_if_no_pyfftw
-def test_pyfftw_call_bad_input(direction):
-
-    # Complex
-
-    # Bad dtype
-    dtype_in = 'complex128'
-    arr_in = np.empty(3, dtype=dtype_in)
-    bad_dtypes_out = np.sctypes['float'] + np.sctypes['complex']
-    try:
-        # This one is correct, so we remove it
-        bad_dtypes_out.remove(np.dtype('complex128'))
-    except ValueError:
-        pass
-    for bad_dtype in bad_dtypes_out:
-        arr_out = np.empty(3, dtype=bad_dtype)
-        with pytest.raises(ValueError):
-            pyfftw_call(arr_in, arr_out, halfcomplex=False,
-                        direction=direction)
-
-    # Bad shape
-    shape = (3, 4)
-    arr_in = np.empty(shape, dtype='complex128')
-    bad_shapes_out = [(3, 3), (3,), (4,), (3, 4, 5), ()]
-    for bad_shape in bad_shapes_out:
-        arr_out = np.empty(bad_shape, dtype='complex128')
-        with pytest.raises(ValueError):
-            pyfftw_call(arr_in, arr_out, halfcomplex=False,
-                        direction=direction)
-
-    # Duplicate axes
-    arr_in = np.empty((3, 4, 5), dtype='complex128')
-    arr_out = np.empty_like(arr_in)
-    bad_axes_list = [(0, 0, 1), (1, 1, 1), (-1, -1)]
-    for bad_axes in bad_axes_list:
-        with pytest.raises(ValueError):
-            pyfftw_call(arr_in, arr_out, axes=bad_axes,
-                        direction=direction)
-
-    # Axis entry out of range
-    arr_in = np.empty((3, 4, 5), dtype='complex128')
-    arr_out = np.empty_like(arr_in)
-    bad_axes_list = [(0, 3), (-4,)]
-    for bad_axes in bad_axes_list:
-        with pytest.raises(IndexError):
-            pyfftw_call(arr_in, arr_out, axes=bad_axes,
-                        direction=direction)
-
-    # Halfcomplex not possible for complex data
-    arr_in = np.empty((3, 4, 5), dtype='complex128')
-    arr_out = np.empty_like(arr_in)
-    with pytest.raises(ValueError):
-        pyfftw_call(arr_in, arr_out, halfcomplex=True,
-                    direction=direction)
-
-    # Data type mismatch
-    arr_in = np.empty((3, 4, 5), dtype='complex128')
-    arr_out = np.empty_like(arr_in, dtype='complex64')
-    with pytest.raises(ValueError):
-        pyfftw_call(arr_in, arr_out, direction=direction)
-
-    # Halfcomplex
-
-    # Bad dtype
-    dtype_in = 'float64'
-    arr_in = np.empty(10, dtype=dtype_in)
-    bad_dtypes_out = np.sctypes['float'] + np.sctypes['complex']
-    try:
-        # This one is correct, so we remove it
-        bad_dtypes_out.remove(np.dtype('complex128'))
-    except ValueError:
-        pass
-    for bad_dtype in bad_dtypes_out:
-        arr_out = np.empty(6, dtype=bad_dtype)
-        with pytest.raises(ValueError):
-            if direction == 'forward':
-                pyfftw_call(arr_in, arr_out, halfcomplex=True,
-                            direction='forward')
-            else:
-                pyfftw_call(arr_out, arr_in, halfcomplex=True,
-                            direction='backward')
-
-    # Bad shape
-    shape = (3, 4, 5)
-    axes_list = [None, (0, 1), (1,), (1, 2), (2, 1), (-1, -2, -3)]
-    arr_in = np.empty(shape, dtype='float64')
-    # Correct shapes:
-    # [(3, 4, 3), (3, 3, 5), (3, 3, 5), (3, 4, 3), (3, 3, 5), (2, 4, 5)]
-    bad_shapes_out = [(3, 4, 2), (3, 4, 3), (2, 3, 5), (3, 2, 3),
-                      (3, 4, 3), (3, 4, 3)]
-    always_bad_shapes = [(3, 4), (3, 4, 5)]
-    for bad_shape, axes in zip(bad_shapes_out, axes_list):
-
-        for always_bad_shape in always_bad_shapes:
-            arr_out = np.empty(always_bad_shape, dtype='complex128')
-            with pytest.raises(ValueError):
-                if direction == 'forward':
-                    pyfftw_call(arr_in, arr_out, axes=axes, halfcomplex=True,
-                                direction='forward')
-                else:
-                    pyfftw_call(arr_out, arr_in, axes=axes, halfcomplex=True,
-                                direction='backward')
-
-        arr_out = np.empty(bad_shape, dtype='complex128')
-        with pytest.raises(ValueError):
-            if direction == 'forward':
-                pyfftw_call(arr_in, arr_out, axes=axes, halfcomplex=True,
-                            direction='forward')
-            else:
-                pyfftw_call(arr_out, arr_in, axes=axes, halfcomplex=True,
-                            direction='backward')
-
-
-@skip_if_no_pyfftw
-def test_pyfftw_call_forward_real_not_halfcomplex():
-    # Test against Numpy's FFT
-    for shape in [(10,), (3, 4, 5)]:
-        arr = _random_array(shape, dtype='float64')
-
-        true_dft = np.fft.fftn(arr)
-        dft_arr = np.empty(shape, dtype='complex128')
-        pyfftw_call(arr, dft_arr, direction='forward', halfcomplex=False)
-
-        assert all_almost_equal(dft_arr, true_dft)
-
-
-@skip_if_no_pyfftw
-def test_pyfftw_call_backward_real_not_halfcomplex():
-    # Test against Numpy's IFFT, no normalization
-    for shape in [(10,), (3, 4, 5)]:
-        # Scaling happens wrt output (large) shape
-        idft_scaling = np.prod(shape)
-
-        arr = _random_array(shape, dtype='float64')
-        true_idft = np.fft.ifftn(arr) * idft_scaling
-        idft_arr = np.empty(shape, dtype='complex128')
-        pyfftw_call(arr, idft_arr, direction='backward', halfcomplex=False)
-
-        assert all_almost_equal(idft_arr, true_idft)
-
-
-@skip_if_no_pyfftw
-def test_pyfftw_call_plan_preserve_input(planning):
-
-    for shape in [(10,), (3, 4)]:
-        arr = _random_array(shape, dtype='complex128')
-        arr_cpy = arr.copy()
-
-        idft_scaling = np.prod(shape)
-        true_idft = np.fft.ifftn(arr) * idft_scaling
-        idft_arr = np.empty(shape, dtype='complex128')
-        pyfftw_call(arr, idft_arr, direction='backward', halfcomplex=False,
-                    planning=planning)
-
-        assert all_almost_equal(arr, arr_cpy)  # Input perserved
-        assert all_almost_equal(idft_arr, true_idft)
-
-        pyfftw_call(arr, idft_arr, direction='backward', halfcomplex=False,
-                    planning=planning)
-
-        assert all_almost_equal(idft_arr, true_idft)
-
-
-@skip_if_no_pyfftw
-def test_pyfftw_call_forward_with_axes(dtype):
-    if dtype == np.dtype('float16'):  # not supported, skipping
-        return
-
-    halfcomplex, out_dtype = _params_from_dtype(dtype)
-    shape = (3, 4, 5)
-
-    test_axes = [(0, 1), [1], (-1,), (1, 0), (-1, -2, -3)]
-    for axes in test_axes:
-        arr = _random_array(shape, dtype)
-        if halfcomplex:
-            true_dft = np.fft.rfftn(arr, axes=axes)
-            dft_arr = np.empty(_halfcomplex_shape(shape, axes),
-                               dtype=out_dtype)
-        else:
-            true_dft = np.fft.fftn(arr, axes=axes)
-            dft_arr = np.empty(shape, dtype=out_dtype)
-
-        pyfftw_call(arr, dft_arr, direction='forward', axes=axes,
-                    halfcomplex=halfcomplex)
-
-        assert all_almost_equal(dft_arr, true_dft)
-
-
-@skip_if_no_pyfftw
-def test_pyfftw_call_backward_with_axes(dtype):
-    if dtype == np.dtype('float16'):  # not supported, skipping
-        return
-
-    halfcomplex, in_dtype = _params_from_dtype(dtype)
-    shape = (3, 4, 5)
-
-    test_axes = [(0, 1), [1], (-1,), (1, 0), (-1, -2, -3)]
-    for axes in test_axes:
-        # Only the shape indexed by axes count for the scaling
-        active_shape = np.take(shape, axes)
-        idft_scaling = np.prod(active_shape)
-
-        if halfcomplex:
-            arr = _random_array(_halfcomplex_shape(shape, axes), in_dtype)
-            true_idft = (np.fft.irfftn(arr, s=active_shape, axes=axes) *
-                         idft_scaling)
-        else:
-            arr = _random_array(shape, in_dtype)
-            true_idft = (np.fft.ifftn(arr, s=active_shape, axes=axes) *
-                         idft_scaling)
-
-        idft_arr = np.empty(shape, dtype=dtype)
-        pyfftw_call(arr, idft_arr, direction='backward', axes=axes,
-                    halfcomplex=halfcomplex)
-
-        assert all_almost_equal(idft_arr, true_idft)
-
-
-@skip_if_no_pyfftw
-def test_pyfftw_call_forward_with_plan():
-
-    for shape in [(10,), (3, 4, 5)]:
-        arr = _random_array(shape, dtype='complex128')
-        arr_cpy = arr.copy()
-        true_dft = np.fft.fftn(arr)
-
-        # First run, create plan
-        dft_arr = np.empty(shape, dtype='complex128')
-        plan = pyfftw_call(arr, dft_arr, direction='forward',
-                           halfcomplex=False, planning_effort='measure')
-
-        # Second run, reuse with fresh output array
-        dft_arr = np.empty(shape, dtype='complex128')
-        pyfftw_call(arr, dft_arr, direction='forward', fftw_plan=plan,
-                    halfcomplex=False)
-
-        assert all_almost_equal(arr, arr_cpy)  # Input perserved
-        assert all_almost_equal(dft_arr, true_dft)
-
-
-@skip_if_no_pyfftw
-def test_pyfftw_call_backward_with_plan():
-
-    for shape in [(10,), (3, 4, 5)]:
-        arr = _random_array(shape, dtype='complex128')
-        arr_cpy = arr.copy()
-        idft_scaling = np.prod(shape)
-        true_idft = np.fft.ifftn(arr) * idft_scaling
-
-        # First run, create plan
-        idft_arr = np.empty(shape, dtype='complex128')
-        plan = pyfftw_call(arr, idft_arr, direction='backward',
-                           halfcomplex=False, planning_effort='measure')
-
-        # Second run, reuse with fresh output array
-        idft_arr = np.empty(shape, dtype='complex128')
-        pyfftw_call(arr, idft_arr, direction='backward', fftw_plan=plan,
-                    halfcomplex=False)
-
-        assert all_almost_equal(arr, arr_cpy)  # Input perserved
-        assert all_almost_equal(idft_arr, true_idft)
-
-
 # ---- DiscreteFourierTransform ---- #
+
+
+impl_params = [never_skip('numpy'), skip_if_no_pyfftw('pyfftw')]
+impl_ids = ['impl={}'.format(impl.args[1]) for impl in impl_params]
+
+
+@pytest.fixture(scope="module", ids=impl_ids, params=impl_params)
+def impl(request):
+    return request.param
 
 
 def test_dft_init(impl):
@@ -1083,6 +335,36 @@ def test_dft_init_plan(impl):
 # ---- FourierTransform ---- #
 
 
+exp_params = [2.0, 1.0, float('inf'), 1.5]
+exp_ids = [' p = {} '.format(p) for p in exp_params]
+
+
+@pytest.fixture(scope="module", ids=exp_ids, params=exp_params)
+def exponent(request):
+    return request.param
+
+
+def _params_from_dtype(dt):
+    if is_real_dtype(dt):
+        halfcomplex = True
+        dtype = TYPE_MAP_R2C[np.dtype(dt)]
+    else:
+        halfcomplex = False
+        dtype = dt
+    return halfcomplex, dtype
+
+
+dtype_params = [str(dtype) for dtype in TYPE_MAP_R2C.keys()]
+dtype_params += [str(dtype) for dtype in TYPE_MAP_R2C.values()]
+dtype_params = list(set(dtype_params))
+dtype_ids = [' dtype = {} '.format(dt) for dt in dtype_params]
+
+
+@pytest.fixture(scope="module", ids=dtype_ids, params=dtype_params)
+def dtype(request):
+    return request.param
+
+
 def test_fourier_trafo_range(exponent, dtype):
     # Check if the range is initialized correctly. Encompasses the init test
 
@@ -1307,6 +589,15 @@ def test_fourier_trafo_sign(impl):
         FourierTransform(discr, sign='+', impl=impl, halfcomplex=True)
     with pytest.raises(ValueError):
         FourierTransform(discr, sign=-1, impl=impl)
+
+
+sign_params = ['-', '+']
+sign_ids = [" sign='{}' ".format(p) for p in sign_params]
+
+
+@pytest.fixture(scope='module', ids=sign_ids, params=sign_params)
+def sign(request):
+    return request.param
 
 
 def test_fourier_trafo_inverse(impl, sign):
